+++ mongo/src/main/kotlin/io/bkbn/lerasium/mongo/Unique.kt
@@ -1,5 +1,0 @@
-package io.bkbn.lerasium.mongo
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.PROPERTY)
-annotation class Unique