+++ playground/src/main/kotlin/io/bkbn/lerasium/playground/domain/BookDomain.kt
@@ -2,7 +2,7 @@ package io.bkbn.lerasium.playground.domain
 
 import io.bkbn.lerasium.api.Api
 import io.bkbn.lerasium.core.Domain
-import io.bkbn.lerasium.rdbms.Index
+import io.bkbn.lerasium.persistence.Index
 import io.bkbn.lerasium.rdbms.Table
 
 @Domain("Book")