+++ playground/src/main/kotlin/io/bkbn/lerasium/playground/domain/UserDomain.kt
@@ -3,8 +3,8 @@ package io.bkbn.lerasium.playground.domain
 import io.bkbn.lerasium.api.Api
 import io.bkbn.lerasium.core.Domain
 import io.bkbn.lerasium.core.Sensitive
-import io.bkbn.lerasium.rdbms.CompositeIndex
-import io.bkbn.lerasium.rdbms.Index
+import io.bkbn.lerasium.persistence.CompositeIndex
+import io.bkbn.lerasium.persistence.Index
 import io.bkbn.lerasium.rdbms.Table
 
 @Domain("User")