+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Count All functionality
 - Get All functionality
 - Set indices on relational tables
+- Set indices on documents 
+- Persistence Annotation Module
 
 ### Changed